requirements.txt
Flask
ffmpeg-python
Werkzeug